# specs/002-github-app/contracts/api.openapi.yaml
openapi: 3.0.3
info:
  title: RepoWeaver GitHub App API
  description: REST API for RepoWeaver GitHub App web interface and webhook handling
  version: 1.0.0
  contact:
    name: RepoWeaver
    url: https://github.com/pradeepmouli/repoweaver
servers:
  - url: https://app.repoweaver.dev
    description: Production
  - url: http://localhost:3000
    description: Local development
security:
  - sessionToken: []
  - githubWebhookSignature: []
paths:
...
  /api/installations:
    get:
      summary: List user's GitHub App installations
      tags: [Installations]
      responses:
        '200':
          description: List of installations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Installation'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/repositories:
    get:
      summary: List repositories accessible to the app
      tags: [Repositories]
      parameters:
        - name: installation_id
          in: query
          schema:
            type: integer
          description: Filter by installation ID
      responses:
        '200':
          description: List of repositories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Repository'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/jobs/{jobId}:
    get:
      summary: Get background job status
      tags: [Jobs]
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Job status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BackgroundJob'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    sessionToken:
      type: apiKey
      in: cookie
      name: session
    githubWebhookSignature:
      type: apiKey
      in: header
      name: X-Hub-Signature-256
  schemas:
    Installation:
      type: object
      properties:
        id:
          type: integer
        github_installation_id:
          type: integer
        account_login:
          type: string
        account_type:
          type: string
          enum: [user, organization]
        installed_at:
          type: integer
          format: int64
    Repository:
      type: object
      properties:
        id:
          type: integer
        full_name:
          type: string
          example: owner/repo-name
        private:
          type: boolean
        has_config:
          type: boolean
          description: Whether RepoWeaver config exists
        auto_update:
          type: boolean
...
    BackgroundJob:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          example: apply_templates
        status:
          type: string
          enum: [pending, running, completed, failed]
        attempts:
          type: integer
        error_message:
          type: string
        created_at:
          type: integer
          format: int64
        completed_at:
          type: integer
          format: int64
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
...
  responses:
    Unauthorized:
      description: Unauthorized - invalid or missing session
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Unauthorized
            message: Please log in to continue
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Not Found
            message: Repository not found